Progress on sub_80039E4

+++ include/gba/defines.h
@@ -61,7 +61,7 @@
 #define TILE_SIZE_8BPP 64
 
 // NOTE/TODO: Maybe this should somewhere else?
-#define GET_TILE_NUM_COMMON(vramPtr, tileSize) (((u32)(vramPtr) - OBJ_VRAM0) / (tileSize))
+#define GET_TILE_NUM_COMMON(vramPtr, tileSize) (((size_t)(vramPtr) - OBJ_VRAM0) / (tileSize))
 #define GET_TILE_NUM(vramPtr) GET_TILE_NUM_COMMON((vramPtr), TILE_SIZE_4BPP) 
 
 #define TOTAL_OBJ_TILE_COUNT 1024
